QuickFit3/imFCS: overview image is double now (not uint16_t) QuickFit3/imFCS: Radhard2 code now omits first channel LIB/Qt/JKQtPlotter: check whether the lines handed over to QPainter::drawLines() are not empty
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@1499 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/interfaces/qfrdrimagetoruninterface.h
@@ -24,7 +24,7 @@ class QFRDRImageToRunInterface {
 
             \return  a uint16_t image of size getDataImageWidth() * getDataImageHeight(), or NULL if no image is available
          */
-        virtual uint16_t* getImageFromRunsPreview() const=0;
+        virtual double* getImageFromRunsPreview() const=0;
         /** \brief convert a pixel coordinate to a rund index */
         virtual int xyToRun(int x, int y) const=0;
         /** \brief convert a run to a pixel x-coordinate */
